requirements.txt
langgraph>=0.0.40
langchain>=0.1.0
langchain-openai>=0.0.8
httpx>=0.25.0
GitPython>=3.1.40
jira>=3.5.0
pytest>=7.4.0
pytest-asyncio>=0.21.0
playwright>=1.40.0
python-dotenv>=1.0.0
pydantic>=2.4.0
fastapi>=0.104.0
uvicorn>=0.24.0
aiofiles>=23.2.0
structlog>=23.1.0
GitPython>=3.1.40
requests>=2.31.0
GitPython>=3.1.40
requests>=2.31.0
python-dotenv
jira>=3.5.0
